Enemy will now attack players when they enter range
Enemies will be 'asleep' until a target (player) walks into their agro trigger and are added to the list of possible targets.

at set intervals the enemy will attempt to attack the nearest player and will keep doing so until there are no more possible targets to attack.

Note:
OnDestroy all invokes for Enemy are canceled but im not 100% sure this is required, but its better to have it in and not need it than having a bunch of invokes running that do nothing